requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
requests>=2.31.0
python-decouple>=3.8
PyMySQL>=1.1.0
django-cors-headers>=4.3.0
django-jazzmin>=2.6.0
django-redis>=5.4.0
Pillow>=10.0.0

